functions/requirements.txt
azure-core==1.30.1
azure-eventgrid==4.19.0
shodan==1.31.0
telebot